refactor(sun): the POWER cache leaves the handler
power_cache_dir, power_cell_key, cached_power_series and the radiation step
were in terra/actions.py, which is where the reader they belong to was not.
They become terra/sun/cache.py.

The cache key is the pair of cells a point resolves to on both grids POWER
serves, because 0.625 does not divide 1.0 and two points inside one meteorology
cell can straddle a radiation boundary. That is a fact about the source, so it
sits beside the module that reads the source. Five actions key on it and four
have no wind in them, which is what made the old arrangement -- the cache
reaching into the wind product for the grid -- visible in the first place.

155 lines out of terra/actions.py, which is now 3,789.

Verification: pytest 518 passed, 8 skipped. ruff check clean, three contracts
kept, gofmt, go vet and go test -race clean. A ping through the entry point
answers.
