+++ sidecar/terra/actions.py
@@ -47,7 +47,6 @@ import json
 import sys
 import warnings
 import xml.etree.ElementTree as ET
-from datetime import UTC, datetime
 from pathlib import Path
 
 import joblib
@@ -868,161 +867,14 @@ def classify_prithvi(products, polygon, ref_profile, model_dir, mode):
 
 # --- Main ------------------------------------------------------------------
 
-def power_cache_dir(req):
-    """
-    Directory the NASA POWER series are cached in, or None if it cannot be made.
-
-    Deliberately outside work_dir: the Go runner creates a fresh temporary
-    work_dir per run, so a cache written under it would never be read by the
-    next one and the fetch would be paid again on every action.
-    """
-    raw = req.get('power_cache_dir')
-    path = Path(raw) if raw else Path.home() / '.cache' / 'geosense' / 'power'
-    try:
-        path.mkdir(parents=True, exist_ok=True)
-    except OSError:
-        return None
-    return path
 
 
-# NASA POWER serves the radiation parameters on a 1 degree grid; see
-# sun_power.GRID_NOTE, which states it beside every solar response.
-POWER_RADIATION_STEP_DEG = 1.0
 
 
-def power_cell_key(lon, lat):
-    """
-    The pair of grid cells a POWER point request resolves to, as a cache key.
-
-    A POWER response is determined by the cell the point falls in on BOTH grids
-    it serves: radiation on 1 degree (sun_power.GRID_NOTE) and meteorology on the
-    MERRA-2 0.5 by 0.625 degree grid (sun_power.meteorology_cell). Keying on
-    alone does not hold, because 0.625 does not divide 1.0, so two points inside
-    one MERRA-2 longitude cell can straddle a radiation cell boundary. The key
-    is the pair, which is the conservative intersection.
-
-    sun_power.request_point rounds to 0.01 degrees, about 1 km and far finer
-    than either grid. Keying the cache on that produced a miss for two AOIs
-    that resolve to identical series and each paid the fetch, so the reuse the
-    cache states it guarantees did not hold.
-    """
-    from terra.sun import nasa_power as sun_power
-
-    met_lon, met_lat = sun_power.meteorology_cell(lon, lat)
-    rad_lon = round(round(float(lon) / POWER_RADIATION_STEP_DEG)
-                    * POWER_RADIATION_STEP_DEG, 6)
-    rad_lat = round(round(float(lat) / POWER_RADIATION_STEP_DEG)
-                    * POWER_RADIATION_STEP_DEG, 6)
-    return f'{met_lon:g}_{met_lat:g}_r{rad_lon:g}_{rad_lat:g}'
 
 
-def cached_power_series(cache_dir, product, lon, lat, start, end, params,
-                        fetch_fn, progress=None):
-    """
-    A POWER series read from disk when one covering `params` is stored, fetched
-    and stored otherwise, with the provenance of whichever path was taken.
-
-    The hourly request is the dominant cost of the whole analysis, measured at
-    about 23 s for ten years, and three actions ask for the same series over
-    the same cell. What determines a POWER response is the
-    grid cell, the period and the parameter list, so those are the key; see
-    power_cell_key for the cell.
-
-    A stored file is reused when its columns cover the requested parameters, so
-    a superset written by one action serves the subset another asks for.
-
-    THE FETCH DATE TRAVELS WITH THE SERIES. POWER reprocesses historical data,
-    so a cached series can be a superseded revision of the record. The cache has
-    no expiry by design, because a research figure benchmarked against a stored
-    run has to stay reproducible; what would make that dangerous is the run not
-    saying so. Each series is stored with the timestamp it was fetched at, and
-    the record returned here is carried into the response of every action that
-    read one, so a cached run and a fetched run are distinguishable.
-
-    Returns (frame, provenance).
-    """
-    import hashlib
-
-    import pandas as pd
-
-    def _record(source, fetched_utc, path=None):
-        return {
-            'source': source,
-            'fetched_utc': fetched_utc,
-            'product': product,
-            'cell_key': cell,
-            'period': f'{start}-{end}',
-            'cache_file': None if path is None else path.name,
-            'note': (
-                'Fetched from NASA POWER during this run.'
-                if source == 'fetch' else
-                'Read from the on-disk POWER cache, not fetched during this '
-                'run. POWER reprocesses historical data, so a series fetched '
-                'earlier can be a superseded revision of the record; the fetch '
-                'timestamp above is when it was retrieved.'
-                if source == 'cache' else
-                'No cache directory was available, so the series was fetched '
-                'and not stored.'
-            ),
-        }
-
-    def now():
-        return datetime.now(UTC).replace(microsecond=0).isoformat()
-    cell = power_cell_key(lon, lat)
-
-    if cache_dir is None:
-        return fetch_fn(progress), _record('fetch_uncached', now())
-
-    prefix = f'{product}_{cell}_{start}_{end}_'
-    wanted = set(params)
-    for path in sorted(cache_dir.glob(prefix + '*.parquet')):
-        try:
-            stored = pd.read_parquet(path)
-        except Exception:
-            continue
-        if wanted.issubset(stored.columns):
-            return stored, _record('cache', _cache_fetch_time(path), path)
-
-    df = fetch_fn(progress)
-    fetched = now()
-    key = hashlib.sha1(','.join(sorted(params)).encode()).hexdigest()[:12]
-    final = cache_dir / (prefix + key + '.parquet')
-    partial = cache_dir / (prefix + key + '.parquet.partial')
-    try:
-        df.to_parquet(partial)
-        partial.replace(final)
-        # Written after the parquet, so a stamp never exists for a series that
-        # is not there. A missing stamp reads as an unknown fetch date rather
-        # than as a fresh one.
-        final.with_name(final.name + '.json').write_text(
-            json.dumps({'fetched_utc': fetched, 'product': product,
-                        'cell_key': cell, 'period': f'{start}-{end}',
-                        'params': sorted(params)})
-        )
-    except Exception:
-        # A cache that cannot be written must not fail a run that already holds
-        # its data. The next run simply pays the fetch again.
-        try:
-            partial.unlink()
-        except OSError:
-            pass
-    return df, _record('fetch', fetched, final)
 
 
-def _cache_fetch_time(path):
-    """
-    When a cached series was fetched, from its stamp, or None if unrecorded.
-
-    Files written before the stamp existed have none. Reporting None is the
-    honest answer: the file modification time is when the parquet was written
-    to this disk, which a copy or a restore changes, so it is not the fetch
-    date and must not be presented as one.
-    """
-    stamp = path.with_name(path.name + '.json')
-    try:
-        return json.loads(stamp.read_text()).get('fetched_utc')
-    except Exception:
-        return None
 
 
 SITING_STAGES = ('dem', 'slope', 'cover', 'classes')
@@ -1398,6 +1250,7 @@ def action_canopy_from_aoi(req, work_dir):
     if req.get('class_stats'):
         try:
             from terra.canopy import species as crop_species
+            from terra.sun import cache as power_cache
             suggestion = crop_species.suggest(req['class_stats'])
         except Exception as e:
             suggestion = {'species': None, 'why': f'suggestion failed: {e}'}
@@ -1588,8 +1441,8 @@ def action_canopy_from_aoi(req, work_dir):
             years = int(req.get('hourly_years', 3))
             start = f'{last_year - years + 1}0101'
             end = f'{last_year}1231'
-            hourly, provenance = cached_power_series(
-                power_cache_dir(req),
+            hourly, provenance = power_cache.cached_power_series(
+                power_cache.power_cache_dir(req),
                 'hourly', cell_lon, cell_lat, start, end,
                 sun_power.HOURLY_PARAMS,
                 lambda progress: sun_power.fetch(
@@ -1819,6 +1672,7 @@ def action_solar_resource(req, work_dir):
 
     from terra.energy import pv as pv_mod
     from terra.sun import (
+        cache as power_cache,
         nasa_power as sun_power,
         position as sun_position,
         record as sun_record,
@@ -1844,10 +1698,10 @@ def action_solar_resource(req, work_dir):
     hourly_start = f'{last_year - hourly_years + 1}0101'
     hourly_end = f'{last_year}1231'
 
-    cache = power_cache_dir(req)
+    cache = power_cache.power_cache_dir(req)
     protocol.emit_progress(5, f'NASA POWER daily, {clim_years} years')
     try:
-        daily, daily_provenance = cached_power_series(
+        daily, daily_provenance = power_cache.cached_power_series(
             cache, 'daily', lon, lat, clim_start, clim_end,
             sun_power.DAILY_PARAMS,
             lambda progress: sun_power.fetch(
@@ -1882,7 +1736,7 @@ def action_solar_resource(req, work_dir):
 
     protocol.emit_progress(42, f'NASA POWER hourly, {hourly_years} years')
     try:
-        hourly, hourly_provenance = cached_power_series(
+        hourly, hourly_provenance = power_cache.cached_power_series(
             cache, 'hourly', lon, lat, hourly_start, hourly_end,
             sun_power.HOURLY_PARAMS,
             lambda progress: sun_power.fetch(
@@ -1986,6 +1840,7 @@ def action_solar_terrain(req, work_dir):
     )
     from terra.imagery import composite as comp
     from terra.sun import (
+        cache as power_cache,
         nasa_power as sun_power,
         position as sun_position,
         record as sun_record,
@@ -2052,8 +1907,8 @@ def action_solar_terrain(req, work_dir):
     hourly_end = f'{last_year}1231'
     protocol.emit_progress(28, f'NASA POWER hourly, {hourly_years} years')
     try:
-        hourly, hourly_provenance = cached_power_series(
-            power_cache_dir(req), 'hourly', lon, lat,
+        hourly, hourly_provenance = power_cache.cached_power_series(
+            power_cache.power_cache_dir(req), 'hourly', lon, lat,
             hourly_start, hourly_end, sun_power.HOURLY_PARAMS,
             lambda progress: sun_power.fetch(
                 'hourly', lon, lat, hourly_start, hourly_end,
@@ -2328,6 +2183,7 @@ def action_energy_model(req, work_dir):
 
     from terra.energy import pv as pv_mod, pv_plant as energy_mod, siting as siting_mod
     from terra.sun import (
+        cache as power_cache,
         nasa_power as sun_power,
         position as sun_position,
         record as sun_record,
@@ -2405,10 +2261,10 @@ def action_energy_model(req, work_dir):
     clim_window = f'{last_year - clim_years + 1}-{last_year} daily'
     hourly_window = f'{last_year - hourly_years + 1}-{last_year} hourly'
 
-    cache = power_cache_dir(req)
+    cache = power_cache.power_cache_dir(req)
     protocol.emit_progress(4, f'NASA POWER daily, {clim_years} years')
     try:
-        daily, daily_provenance = cached_power_series(
+        daily, daily_provenance = power_cache.cached_power_series(
             cache, 'daily', lon, lat, clim_start, clim_end,
             sun_power.DAILY_PARAMS,
             lambda progress: sun_power.fetch(
@@ -2427,7 +2283,7 @@ def action_energy_model(req, work_dir):
 
     protocol.emit_progress(32, f'NASA POWER hourly, {hourly_years} years')
     try:
-        hourly, hourly_provenance = cached_power_series(
+        hourly, hourly_provenance = power_cache.cached_power_series(
             cache, 'hourly', lon, lat, hourly_start, hourly_end,
             sun_power.HOURLY_PARAMS,
             lambda progress: sun_power.fetch(
@@ -2661,7 +2517,7 @@ def action_wind_resource(req, work_dir):
     from datetime import date as _date
 
     from terra.energy import wind as wind_mod
-    from terra.sun import nasa_power as sun_power
+    from terra.sun import cache as power_cache, nasa_power as sun_power
 
     if not req.get('polygon_geojson'):
         protocol.fail('no polygon provided (polygon_geojson required)')
@@ -2700,8 +2556,8 @@ def action_wind_resource(req, work_dir):
 
     protocol.emit_progress(5, f'NASA POWER hourly wind, {record_years} years')
     try:
-        df, hourly_provenance = cached_power_series(
-            power_cache_dir(req), 'hourly', lon, lat, start, end,
+        df, hourly_provenance = power_cache.cached_power_series(
+            power_cache.power_cache_dir(req), 'hourly', lon, lat, start, end,
             wind_mod.HOURLY_PARAMS,
             lambda progress: wind_mod.fetch(
                 lon, lat, start, end, progress=progress
